some changes
`pmscore::arduino::encoder` incrementation is now done by using interrupt
routines.

`pmscore::correcter` uses now two different speed profiles, one for the robot
speed and the other for its rotation speed.

Add debug macros.

Fix `angle_distance`.

`pmscore::speed_profile::triangular` is now more generalized and is renamed to
`pmscore::speed_progile::trapezoidal`.

+++ pmscore/src/arduino/motoreductor.cpp
@@ -86,10 +86,10 @@ namespace pmscore::arduino
 
         if (m_power < 0) {
             set_direction(direction::Back);
-            analogWrite(m_pin_pwm, -m_power);
+            analogWrite(m_pin_pwm, static_cast<int>(-m_power));
         } else {
             set_direction(direction::Front);
-            analogWrite(m_pin_pwm, m_power);
+            analogWrite(m_pin_pwm, static_cast<int>(m_power));
         }
     }
 }
